+++ contracts/src/mocks/plugin/PluginCloneableSetupMock.sol
@@ -6,6 +6,7 @@ pragma solidity ^0.8.8;
 import {PermissionLib} from "../../permission/PermissionLib.sol";
 import {IPluginSetup} from "../../plugin/setup/IPluginSetup.sol";
 import {PluginSetup} from "../../plugin/setup/PluginSetup.sol";
+import {ProxyLib} from "../../utils/deployment/ProxyLib.sol";
 import {IDAO} from "../../dao/IDAO.sol";
 import {mockPermissions, mockHelpers} from "./PluginSetupMockData.sol";
 import {PluginCloneableMockBuild1, PluginCloneableMockBuild2} from "./PluginCloneableMock.sol";
@@ -14,6 +15,8 @@ import {PluginCloneableMockBuild1, PluginCloneableMockBuild2} from "./PluginClon
 /// v1.1 (Release 1, Build 1)
 /// @dev DO NOT USE IN PRODUCTION!
 contract PluginCloneableSetupMockBuild1 is PluginSetup {
+    using ProxyLib for address;
+
     address internal pluginBase;
 
     constructor() {
@@ -24,9 +27,9 @@ contract PluginCloneableSetupMockBuild1 is PluginSetup {
     function prepareInstallation(
         address _dao,
         bytes memory
-    ) external override returns (address plugin, PreparedSetupData memory preparedSetupData) {
+    ) external returns (address plugin, PreparedSetupData memory preparedSetupData) {
         bytes memory initData = abi.encodeCall(PluginCloneableMockBuild1.initialize, (IDAO(_dao)));
-        plugin = createERC1967Proxy(pluginBase, initData); // TODO createClone(pluginBase, initData); is missing! See task OS-794 and OS-675.
+        plugin = pluginBase.deployMinimalProxy(initData);
         preparedSetupData.helpers = mockHelpers(1);
         preparedSetupData.permissions = mockPermissions(0, 1, PermissionLib.Operation.Grant);
     }
@@ -35,13 +38,13 @@ contract PluginCloneableSetupMockBuild1 is PluginSetup {
     function prepareUninstallation(
         address _dao,
         SetupPayload calldata _payload
-    ) external pure override returns (PermissionLib.MultiTargetPermission[] memory permissions) {
+    ) external pure returns (PermissionLib.MultiTargetPermission[] memory permissions) {
         (_dao, _payload);
         permissions = mockPermissions(0, 1, PermissionLib.Operation.Revoke);
     }
 
     /// @inheritdoc IPluginSetup
-    function implementation() external view override returns (address) {
+    function implementation() external view returns (address) {
         return address(pluginBase);
     }
 }
@@ -50,6 +53,8 @@ contract PluginCloneableSetupMockBuild1 is PluginSetup {
 /// v1.2 (Release 1, Build 2)
 /// @dev DO NOT USE IN PRODUCTION!
 contract PluginCloneableSetupMockBuild2 is PluginSetup {
+    using ProxyLib for address;
+
     address internal pluginBase;
 
     constructor() {
@@ -60,9 +65,9 @@ contract PluginCloneableSetupMockBuild2 is PluginSetup {
     function prepareInstallation(
         address _dao,
         bytes memory
-    ) external override returns (address plugin, PreparedSetupData memory preparedSetupData) {
+    ) external returns (address plugin, PreparedSetupData memory preparedSetupData) {
         bytes memory initData = abi.encodeCall(PluginCloneableMockBuild2.initialize, (IDAO(_dao)));
-        plugin = createERC1967Proxy(pluginBase, initData); // TODO createClone(pluginBase, initData); is missing! See task OS-794 and OS-675.
+        plugin = pluginBase.deployMinimalProxy(initData);
         preparedSetupData.helpers = mockHelpers(2);
         preparedSetupData.permissions = mockPermissions(0, 2, PermissionLib.Operation.Grant);
     }
@@ -71,13 +76,13 @@ contract PluginCloneableSetupMockBuild2 is PluginSetup {
     function prepareUninstallation(
         address _dao,
         SetupPayload calldata _payload
-    ) external pure override returns (PermissionLib.MultiTargetPermission[] memory permissions) {
+    ) external pure returns (PermissionLib.MultiTargetPermission[] memory permissions) {
         (_dao, _payload);
         permissions = mockPermissions(0, 2, PermissionLib.Operation.Revoke);
     }
 
     /// @inheritdoc IPluginSetup
-    function implementation() external view override returns (address) {
+    function implementation() external view returns (address) {
         return address(pluginBase);
     }
 }