+++ contracts/src/mocks/plugin/PluginUUPSUpgradeableSetupMock.sol
@@ -5,7 +5,8 @@ pragma solidity ^0.8.8;
 
 import {PermissionLib} from "../../permission/PermissionLib.sol";
 import {IPluginSetup} from "../../plugin/setup/IPluginSetup.sol";
-import {PluginSetup} from "../../plugin/setup/PluginSetup.sol";
+import {PluginUUPSUpgradeableSetup} from "../../plugin/setup/PluginUUPSUpgradeableSetup.sol";
+import {ProxyLib} from "../../utils/deployment/ProxyLib.sol";
 import {IDAO} from "../../dao/IDAO.sol";
 import {mockPermissions, mockHelpers} from "./PluginSetupMockData.sol";
 import {PluginUUPSUpgradeableMockBuild1, PluginUUPSUpgradeableMockBuild2, PluginUUPSUpgradeableMockBuild3} from "./PluginUUPSUpgradeableMock.sol";
@@ -13,7 +14,9 @@ import {PluginUUPSUpgradeableMockBuild1, PluginUUPSUpgradeableMockBuild2, Plugin
 /// @notice A mock plugin setup of an upgradeable plugin to be deployed via the UUPS pattern.
 /// v1.1 (Release 1, Build 1)
 /// @dev DO NOT USE IN PRODUCTION!
-contract PluginUUPSUpgradeableSetupMockBuild1 is PluginSetup {
+contract PluginUUPSUpgradeableSetupMockBuild1 is PluginUUPSUpgradeableSetup {
+    using ProxyLib for address;
+
     address internal pluginBase;
 
     constructor() {
@@ -24,27 +27,38 @@ contract PluginUUPSUpgradeableSetupMockBuild1 is PluginSetup {
     function prepareInstallation(
         address _dao,
         bytes memory
-    ) public virtual override returns (address plugin, PreparedSetupData memory preparedSetupData) {
+    ) public returns (address plugin, PreparedSetupData memory preparedSetupData) {
         bytes memory initData = abi.encodeCall(
             PluginUUPSUpgradeableMockBuild1.initialize,
             (IDAO(_dao))
         );
-        plugin = createERC1967Proxy(pluginBase, initData);
+        plugin = pluginBase.deployUUPSProxy(initData);
         preparedSetupData.helpers = mockHelpers(1);
         preparedSetupData.permissions = mockPermissions(0, 1, PermissionLib.Operation.Grant);
+    }
+
+    /// @inheritdoc IPluginSetup
+    /// @dev Build 1 is the first build so this function reverts.
+    function prepareUpdate(
+        address _dao,
+        uint16 _fromBuild,
+        SetupPayload calldata _payload
+    ) external pure returns (bytes memory, PreparedSetupData memory) {
+        (_dao, _fromBuild, _payload);
+        revert InvalidUpdatePath({fromBuild: 0, thisBuild: 1});
     }
 
     /// @inheritdoc IPluginSetup
     function prepareUninstallation(
         address _dao,
         SetupPayload calldata _payload
-    ) external virtual override returns (PermissionLib.MultiTargetPermission[] memory permissions) {
+    ) external pure returns (PermissionLib.MultiTargetPermission[] memory permissions) {
         (_dao, _payload);
         permissions = mockPermissions(0, 1, PermissionLib.Operation.Revoke);
     }
 
     /// @inheritdoc IPluginSetup
-    function implementation() external view virtual override returns (address) {
+    function implementation() external view returns (address) {
         return address(pluginBase);
     }
 }
@@ -52,7 +66,9 @@ contract PluginUUPSUpgradeableSetupMockBuild1 is PluginSetup {
 /// @notice A mock plugin setup of an upgradeable plugin to be deployed via the UUPS pattern.
 /// v1.2 (Release 1, Build 2)
 /// @dev DO NOT USE IN PRODUCTION!
-contract PluginUUPSUpgradeableSetupMockBuild2 is PluginSetup {
+contract PluginUUPSUpgradeableSetupMockBuild2 is PluginUUPSUpgradeableSetup {
+    using ProxyLib for address;
+
     address internal pluginBase;
 
     constructor() {
@@ -63,12 +79,12 @@ contract PluginUUPSUpgradeableSetupMockBuild2 is PluginSetup {
     function prepareInstallation(
         address _dao,
         bytes memory
-    ) external override returns (address plugin, PreparedSetupData memory preparedSetupData) {
+    ) external returns (address plugin, PreparedSetupData memory preparedSetupData) {
         bytes memory initData = abi.encodeCall(
             PluginUUPSUpgradeableMockBuild2.initialize,
             (IDAO(_dao))
         );
-        plugin = createERC1967Proxy(pluginBase, initData);
+        plugin = pluginBase.deployUUPSProxy(initData);
         preparedSetupData.helpers = mockHelpers(2);
         preparedSetupData.permissions = mockPermissions(0, 2, PermissionLib.Operation.Grant);
     }
@@ -76,23 +92,15 @@ contract PluginUUPSUpgradeableSetupMockBuild2 is PluginSetup {
     /// @inheritdoc IPluginSetup
     function prepareUpdate(
         address _dao,
-        uint16 _currentBuild,
+        uint16 _fromBuild,
         SetupPayload calldata _payload
-    )
-        external
-        pure
-        override
-        returns (bytes memory initData, PreparedSetupData memory preparedSetupData)
-    {
+    ) external pure returns (bytes memory initData, PreparedSetupData memory preparedSetupData) {
         (_dao, _payload);
 
         // Update from Build 1
-        if (_currentBuild == 1) {
+        if (_fromBuild == 1) {
             preparedSetupData.helpers = mockHelpers(2);
-            initData = abi.encodeCall(
-                PluginUUPSUpgradeableMockBuild2.initializeFrom,
-                (_currentBuild)
-            );
+            initData = abi.encodeCall(PluginUUPSUpgradeableMockBuild2.initializeFrom, (_fromBuild));
             preparedSetupData.permissions = mockPermissions(1, 2, PermissionLib.Operation.Grant);
         }
     }
@@ -101,20 +109,22 @@ contract PluginUUPSUpgradeableSetupMockBuild2 is PluginSetup {
     function prepareUninstallation(
         address _dao,
         SetupPayload calldata _payload
-    ) external pure override returns (PermissionLib.MultiTargetPermission[] memory permissions) {
+    ) external pure returns (PermissionLib.MultiTargetPermission[] memory permissions) {
         (_dao, _payload);
         permissions = mockPermissions(0, 2, PermissionLib.Operation.Revoke);
     }
 
     /// @inheritdoc IPluginSetup
-    function implementation() external view override returns (address) {
+    function implementation() external view returns (address) {
         return address(pluginBase);
     }
 }
 
 /// @notice A mock plugin setup of an upgradeable plugin to be deployed via the UUPS pattern.
 /// v1.3 (Release 1, Build 3)
-contract PluginUUPSUpgradeableSetupMockBuild3 is PluginSetup {
+contract PluginUUPSUpgradeableSetupMockBuild3 is PluginUUPSUpgradeableSetup {
+    using ProxyLib for address;
+
     address internal pluginBase;
 
     constructor() {
@@ -130,7 +140,7 @@ contract PluginUUPSUpgradeableSetupMockBuild3 is PluginSetup {
             PluginUUPSUpgradeableMockBuild3.initialize,
             (IDAO(_dao))
         );
-        plugin = createERC1967Proxy(pluginBase, initData);
+        plugin = pluginBase.deployUUPSProxy(initData);
         preparedSetupData.helpers = mockHelpers(3);
         preparedSetupData.permissions = mockPermissions(0, 3, PermissionLib.Operation.Grant);
     }
@@ -138,33 +148,22 @@ contract PluginUUPSUpgradeableSetupMockBuild3 is PluginSetup {
     /// @inheritdoc IPluginSetup
     function prepareUpdate(
         address _dao,
-        uint16 _currentBuild,
+        uint16 _fromBuild,
         SetupPayload calldata _payload
-    )
-        external
-        pure
-        override
-        returns (bytes memory initData, PreparedSetupData memory preparedSetupData)
-    {
+    ) external pure returns (bytes memory initData, PreparedSetupData memory preparedSetupData) {
         (_dao, _payload);
 
         // Update from Build 1
-        if (_currentBuild == 1) {
+        if (_fromBuild == 1) {
             preparedSetupData.helpers = mockHelpers(3);
-            initData = abi.encodeCall(
-                PluginUUPSUpgradeableMockBuild3.initializeFrom,
-                (_currentBuild)
-            );
+            initData = abi.encodeCall(PluginUUPSUpgradeableMockBuild3.initializeFrom, (_fromBuild));
             preparedSetupData.permissions = mockPermissions(1, 3, PermissionLib.Operation.Grant);
         }
 
         // Update from Build 2
-        if (_currentBuild == 2) {
+        if (_fromBuild == 2) {
             preparedSetupData.helpers = mockHelpers(3);
-            initData = abi.encodeCall(
-                PluginUUPSUpgradeableMockBuild3.initializeFrom,
-                (_currentBuild)
-            );
+            initData = abi.encodeCall(PluginUUPSUpgradeableMockBuild3.initializeFrom, (_fromBuild));
             preparedSetupData.permissions = mockPermissions(2, 3, PermissionLib.Operation.Grant);
         }
     }
@@ -173,13 +172,13 @@ contract PluginUUPSUpgradeableSetupMockBuild3 is PluginSetup {
     function prepareUninstallation(
         address _dao,
         SetupPayload calldata _payload
-    ) external virtual override returns (PermissionLib.MultiTargetPermission[] memory permissions) {
+    ) external pure returns (PermissionLib.MultiTargetPermission[] memory permissions) {
         (_dao, _payload);
         permissions = mockPermissions(0, 3, PermissionLib.Operation.Revoke);
     }
 
     /// @inheritdoc IPluginSetup
-    function implementation() external view override returns (address) {
+    function implementation() external view returns (address) {
         return address(pluginBase);
     }
 }